#6: 회원가입 시 프로필 이미지 파일 업로드 및 불러오기 기능
