+++ sorisam/src/main/java/com/sparta/sorisam/config/AmazonS3Config.java
@@ -25,6 +25,4 @@ public class AmazonS3Config {
                 .withCredentials(new AWSStaticCredentialsProvider(awsCreds))
                 .build();
     }
-
-
 }
