+++ sorisam/src/main/java/com/sparta/sorisam/controller/UserController.java
@@ -8,6 +8,7 @@ import com.sparta.sorisam.global.common.response.CommonResponse;
 import com.sparta.sorisam.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -15,18 +16,20 @@ import javax.validation.Valid;
 
 @RestController
 @RequiredArgsConstructor
+@RequestMapping("/api")
 public class UserController {
 
     private final UserService userService;
 
     @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping("/api/signup")
-    public CommonResponse<?> registerUser(@Valid @RequestBody SignupRequestDto requestDto, MultipartFile img) {
+    @PostMapping(value = "signup",consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
+    public CommonResponse<?> registerUser(@RequestPart @Valid SignupRequestDto requestDto,
+                                          @RequestPart(required = false) MultipartFile img) {
         userService.registerUser(requestDto,img);
         return ApiUtils.success(201, "회원가입에 성공하였습니다.");
     }
 
-    @PostMapping("/api/login")
+    @PostMapping("/login")
     public CommonResponse<?> login(@RequestBody LoginRequestDto loginRequestDto) {
         String token = userService.login(loginRequestDto);
         return ApiUtils.success(200, token);
